khanbuta/requirements.txt
asgiref==3.5.0
# Django==4.0.2
djangorestframework==3.13.1
django-cors-headers==3.11.0
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
